% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pIRRitchard.R
\name{pIRRitchard}
\alias{pIRRitchard}
\title{pIRRitchard}
\usage{
pIRRitchard(data, n_raters, type)
}
\arguments{
\item{data}{a wide data.frame where col 1 = unit ID and subsequent columns are a criteria/variable and rater progression. For 2 variables and2 raters: C1id, C2v1r1, C3v1r2, C4v2r1, C5v2r2.}

\item{n_raters}{number of raters. Used in determining sequences pulled for reliability estimates.}

\item{type}{a string of 'fleiss' or 'ac1'.}
}
\description{
Quickly get reliability estimates of multiple criteria/variables and raters
}
