% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pIRRitchard_data.R
\name{piRRitchard_data}
\alias{piRRitchard_data}
\title{pIRRitchard_data}
\usage{
piRRitchard_data(criteria, raters, units)
}
\arguments{
\item{criteria}{number of criteria/variables to simulate}

\item{raters}{number of raters to simulate}

\item{units}{number os unit to simulate}
}
\value{
A unit X critera*raters data.frame
}
\description{
Create a sample data.frame in wide form to test the function.
}
