% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pIRRitchard_plot.R
\name{pIRRitchard_plot}
\alias{pIRRitchard_plot}
\title{pIRRitchard_plot}
\usage{
pIRRitchard_plot(data, benchmarks = T)
}
\arguments{
\item{data}{an output from the main pIRRitchard function. Otherwise, a data.frame where column 1 = criteria/variable name and column 2 = fleiss/AC1 value}

\item{benchmarks}{TRUE or FALSE. Include shaded boundaries of .2 width.}
}
\description{
Create a histogram of reliability estimates with or without shading boundaries.
}
